Revert "Sprint 3/step 1 (#5)"
This reverts commit a7c7f66491440fff688012857e6fcbbc3304b7b7.
