+++ package.json
@@ -18,7 +18,6 @@
     "react-dnd-html5-backend": "^14.0.0",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
     "redux-thunk": "^2.3.0",
