+++ src/components/app/app.module.scss
@@ -1,0 +1,13 @@
+* {
+    box-sizing: border-box;
+}
+.container {
+    display: flex;
+    max-width: 1280px;
+    padding-left: 20px;
+    padding-right: 20px;
+    margin: 0 auto;
+    @media screen and (max-width: 991px) {
+        flex-direction: column;
+    }
+}
