--- src/components/feed-info/feedInfo.module.scss
@@ -1,18 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    padding-left: 56px;
-    width: calc(100% - 56px);
-}
-.list {
-    list-style-type: none;
-    padding: 0;
-}
-.status {
-    display: flex;
-    width: 100%;
-    div {
-        max-width: 280px;
-        width: 100%;
-    }
-}
