--- src/components/feed-info/index.js
@@ -1,2 +1,0 @@
-import FeedInfo from './feedInfo'
-export {FeedInfo}
