--- src/components/orderFeed/index.js
@@ -1,2 +1,0 @@
-import OrderFeed from './orderFeed'
-export {OrderFeed}
