--- src/components/orderFeed/orderFeed.module.scss
@@ -1,25 +1,0 @@
-.orderFeed {
-    display: flex;
-    max-width: 50%;
-    flex-direction: column;
-    max-height: 80vh;
-    width: 100%;
-    overflow-y: auto;
-    overflow-x: hidden;
-    padding-right: 10px;
-    &:last-child {
-        margin-bottom: 0;
-    }
-    &::-webkit-scrollbar {
-        width: 12px;
-    }
-    &::-webkit-scrollbar-track {
-        background: #2F2F37;
-        margin-bottom: 52px;
-    }
-    &::-webkit-scrollbar-thumb {
-        background: #8585AD;
-        border: 3px solid #2F2F37;
-        border-radius: 6px;
-    }
-}
