--- src/components/orderFeedItem/index.js
@@ -1,2 +1,0 @@
-import OrderFeedItem from './orderFeedItem'
-export {OrderFeedItem}
