--- src/components/orderFeedItem/orderFeedItem.module.scss
@@ -1,72 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    padding: 24px;
-    background: #1C1C21;
-    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 2px rgba(0, 0, 0, 0.06), 0px 0px 1px rgba(0, 0, 0, 0.04);
-    border-radius: 40px;
-    margin-bottom: 10px;
-    &:last-child {
-        margin-bottom: 0;
-    }
-    color: #fff;
-    text-decoration: none;
-}
-.header {
-    display: flex;
-    justify-content: space-between;
-}
-.body {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-}
-.name {
-    width: 100%;
-}
-.price {
-    display: flex;
-    align-items: center;
-    margin-top: 28px;
-    svg {
-        margin-right: 10px;
-    }
-}
-.items {
-    display: flex;
-    min-width: 80%;
-    margin-top: 28px;
-}
-.item {
-    position: relative;
-    background-image: linear-gradient(63.18deg, #801AB3 0%, #4C4CFF 100%);
-    border-radius: 50%;
-    height: 64px;
-    width: 64px;
-    overflow: hidden;
-    img {
-        display: block;
-        position: absolute;
-        top: 2px;
-        left: 2px;
-        height: 60px;
-        width: 60px;
-        border-radius: 50%;
-    }
-    &:first-child {
-        left: 0;
-    }
-}
-.img_opac {
-    opacity: 0.7;
-}
-.placeHolder {
-    display: flex;
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    align-items: center;
-    justify-content: center;
-}
