--- src/components/profile-info/index.js
@@ -1,2 +1,0 @@
-import ProfileInfo from './profileInfo'
-export {ProfileInfo}
