--- src/components/profile-info/profileInfo.module.scss
@@ -1,21 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-
-}
-.link {
-    text-decoration: none;
-}
-.active {
-    color: #fff;
-}
-.input {
-    margin-bottom: 20px;
-}
-.cancel {
-    margin-right: 10px;
-    font-family: JetBrains Mono;
-    font-size: 16px;
-    line-height: 24px;
-    color: #4C4CFF;
-}
