--- src/components/profile-nav/index.js
@@ -1,2 +1,0 @@
-import ProfileNav from './profileNav'
-export {ProfileNav}
