--- src/components/profile-nav/profileNav.module.scss
@@ -1,12 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    list-style-type: none;
-    padding: 0;
-}
-.link {
-    text-decoration: none;
-}
-.active {
-    color: #fff;
-}
