--- src/components/profile-orders/index.js
@@ -1,2 +1,0 @@
-import ProfileOrders from './profileOrders'
-export {ProfileOrders}
