+++ src/index.css
@@ -7,7 +7,3 @@ body {
     height: 100vh;
     padding-bottom: 20px;
 }
-.link {
-    color: initial;
-    text-decoration: none;
-}
