--- src/pages/feed/index.js
@@ -1,2 +1,0 @@
-import Feed from './feed'
-export {Feed}
