--- src/pages/forgot-password/forgot.js
@@ -1,53 +1,0 @@
-import { Button, Input, PasswordInput, Logo } from "@ya.praktikum/react-developer-burger-ui-components";
-import styles from './forgot.module.scss'
-import { Link } from "react-router-dom";
-import React from 'react'
-export default function Forgot () {
-    const [value, setValue] = React.useState({
-        email: '',
-    })
-    const setValueInput = (e) => {
-      setValue((prev) => ({
-        ...prev,
-        [e.target.name]: e.target.value
-      }))
-    }
-    const submit = (e) => {
-        e.preventDefault();
-        fetch('https://norma.nomoreparties.space/api/password-reset', {
-            method: 'POST',
-            body: JSON.stringify(value)
-        }).then( e => {
-            if(e.ok) {
-                return e.json();
-              }
-            return Promise.reject(e)
-        }).then( e => {
-            console.log(e);
-        }).catch( e => console.log(e))
-    }
-    return (
-        <div className={styles.container}>
-            <div className={styles.logo}><Logo /></div>
-            <h1 className={`text text_type_main-medium ${styles.title}`}> Восстановление пароля</h1>
-            <form className={styles.form} onSubmit={submit}>
-                <div className={styles.input}> 
-                    <Input
-                        onChange={setValueInput}
-                        type={'text'}
-                        value={value.email}
-                        name={'email'}
-                        size={'default'}
-                        placeholder={'Укажите e-mail'}
-                    />
-                </div>
-                <Button type="primary" size="medium">
-                    Восстановить
-                </Button>
-            </form>
-          <span className="text text_type_main-default text_color_inactive">
-          Вспомнили пароль? <Link to={'/login'} className={styles.link}>Войти</Link>
-          </span>
-        </div>
-    )
-}
