--- src/pages/forgot-password/index.js
@@ -1,2 +1,0 @@
-import ForgotPassword from './forgot'
-export { ForgotPassword }
