--- src/pages/index.js
@@ -1,9 +1,0 @@
-import { Main } from './main'
-import { Login } from './login'
-import { Registration } from './registration'
-import { ForgotPassword } from './forgot-password';
-import { ResetPassword } from './reset-password';
-import { Feed } from './feed';
-import { Profile } from './profile';
-import { OrderDetails } from './order-details';
-export { Main, Login, Registration, ForgotPassword, ResetPassword, Feed, Profile, OrderDetails}
