--- src/pages/login/index.js
@@ -1,2 +1,0 @@
-import Login from './login'
-export { Login }
