--- src/pages/login/login.module.scss
@@ -1,30 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin: 0 auto;
-    padding-top: 120px;
-    form {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        margin-bottom: 80px;
-    } 
-}
-.link {
-    font-size: 16px;
-    line-height: 24px;
-    color: #4C4CFF;
-    text-decoration: none;
-}
-.logo {
-    margin-bottom: 80px;
-}
-.title {
-    margin-bottom: 24px;
-}
-.input {
-    margin-bottom: 24px;
-}
