--- src/pages/main/index.js
@@ -1,2 +1,0 @@
-import Main from './main.js'
-export { Main };
