--- src/pages/order-details/index.js
@@ -1,2 +1,0 @@
-import OrderDetails from './orderDetails';
-export { OrderDetails }
