--- src/pages/order-details/orderDetails.module.scss
@@ -1,127 +1,0 @@
-.container {
-    max-width: 640px;
-    margin: 0 auto;
-    width: 100%;
-    margin-top: 50px;
-}
-.item {
-    display: flex;
-    align-items: center;
-    margin-bottom: 8px;
-    &:last-child {
-        margin-bottom: 0;
-    }
-}
-.header {
-    display: flex;
-    flex-direction: column;
-    margin-bottom: 60px;
-}
-.name {
-    font-family: JetBrains Mono;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 24px;
-    line-height: 30px;
-    margin-bottom: 8px;
-}
-.num {
-    font-family: Iceland;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 28px;
-    line-height: 24px;
-    align-self: center;
-    margin-bottom: 20px;
-}
-.status {
-    font-family: JetBrains Mono;
-    font-size: 16px;
-    line-height: 24px;
-    color: #00CCCC;
-}
-.img {
-    position: relative;
-    background-image: linear-gradient(63.18deg, #801AB3 0%, #4C4CFF 100%);
-    border-radius: 50%;
-    height: 64px;
-    width: 64px;
-    overflow: hidden;
-    img {
-        display: block;
-        position: absolute;
-        top: 2px;
-        left: 2px;
-        height: 60px;
-        width: 60px;
-        border-radius: 50%;
-    }
-}
-.sostav {
-    font-family: JetBrains Mono;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 24px;
-    line-height: 30px;
-    color: #F2F2F3;
-    margin-bottom: 28px;
-}
-.items {
-    max-height: 360px;
-    overflow-x: hidden;
-    overflow-y: auto;
-    padding-right: 10px;
-    &::-webkit-scrollbar {
-        width: 12px;
-    }
-    &::-webkit-scrollbar-track {
-        background: #2F2F37;
-        margin-bottom: 52px;
-    }
-    &::-webkit-scrollbar-thumb {
-        background: #8585AD;
-        border: 3px solid #2F2F37;
-        border-radius: 6px;
-    }
-}
-.nameitem {
-    margin-left: 16px;
-    margin-right: auto;
-    font-family: JetBrains Mono;
-    font-size: 16px;
-    line-height: 24px;
-    color: #F2F2F3;
-}
-.priceitem {
-    display: flex;
-    font-family: Iceland;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 28px;
-    line-height: 24px;
-    svg {
-        margin-left: 5px;
-    }
-}
-.footer {
-    display: flex;
-    justify-content: space-between;
-    margin-top: 10px;
-}
-.time {
-    font-family: JetBrains Mono;
-    font-size: 16px;
-    line-height: 24px;
-    color: #8585AD;
-}
-.price {
-    display: flex;
-    font-family: Iceland;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 28px;
-    line-height: 24px;
-    svg {
-        margin-left: 5px;
-    }
-}
