--- src/pages/profile/index.js
@@ -1,2 +1,0 @@
-import Profile from './profile'
-export {Profile}
