--- src/pages/profile/profile.module.scss
@@ -1,11 +1,0 @@
-.container {
-    display: flex;
-    padding: 16px 20px;
-    max-width: 1280px;
-    width: 100%;
-    margin: 0 auto;
-    padding-top: 100px;
-}
-.left {
-    width: 30%;
-}
