--- src/pages/registration/index.js
@@ -1,2 +1,0 @@
-import Registration from './registration'
-export { Registration }
