--- src/pages/reset-password/index.js
@@ -1,2 +1,0 @@
-import ResetPassword from './reset'
-export { ResetPassword }
